requirements.txt
# Inference
onnxruntime>=1.16.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Printer API
requests>=2.31.0

# Dashboard
fastapi>=0.104.0
uvicorn>=0.24.0

# Config
python-dotenv>=1.0.0

# Tests
pytest>=7.4.0
pytest-cov>=4.1.0
